requirements.txt
Django==4.2
python-decouple==3.8
psycopg2==2.9.6
djangorestframework==3.14.0
django-cors-headers==4.2.0
PyJWT==2.8.0
djangorestframework-simplejwt==5.2.0
